almost done the case for when flat top of triangle

+++ point2d.h
@@ -3,8 +3,8 @@
 
 struct Point2D
 {
-	float x{};
-	float y{};
+	int x{};
+	int y{};
 };
 
 #endif
